Builds Postgres, darwin-x86_64, but `make check` fails
This commit is the result of a rewrite of the previous attempt that
depended on Nixpkgs` `stdenv`. A rewrite was necessary since the
`stdenv` derivation is very hard to tame; it has a lot of ad-hoc
variables defined, and decisions made. If this project ever hopes to be
adopted by other Postgres community members, then every decision made
should be explainable.

Hence, we now rely on the Nix language as much as possible, use as
little of `stdenv`, as possible, and record all the variables in the
build output, for troubleshooting.