Use Darwin binutils provided by Nixpkgs
We used to use /usr/bin/ar, an impure path, because the `ar` command in
the binutils package used to cause errors. This patch fixes that by
providing not just `ar`, but other commands as well from the
Nixpkgs-supported binutils specifically maintained for Darwin.

This patch includes many other changes as well, mostly cosmetic. But
these are mostly in preparation for the patch we need, to support
`make check`; being discussed in the Postgres community:

https://www.postgresql.org/message-id/CABwTF4VBTKLORbKMOT8qi7W76vQ5hsV73Q%2BraPpXFveoiWUwBQ%40mail.gmail.com